+++ middleware.go
@@ -1,0 +1,19 @@
+package main
+
+import "net/http"
+
+// logMW is a middleware that logs all incoming requests
+func logMW(next http.HandlerFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		log.Debug("request received",
+			"method", r.Method,
+			"remote", r.RemoteAddr,
+			"path", r.URL.Path,
+			"refer", r.Referer(),
+			"user-agent", r.UserAgent(),
+			"opaque", r.URL.Opaque,
+			"bytes", r.ContentLength,
+		)
+		next(w, r)
+	}
+}
